requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
flatbuffers==1.12
google-api-core==1.31.1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-cloud-pubsub==2.7.0
google-cloud-secret-manager==2.6.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.0
pip==21.1.3
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.36.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.5.0
